refactor(thread): clarify runtime architecture
