+++ packages/thread/src/index.ts
@@ -1,16 +1,18 @@
 export {
 	createThreadRuntime,
 	getMessageText,
-	type MessageTreeSnapshot,
-	type MessageTreeStore,
 	ROOT_PARENT_ID,
-	type ThreadConcurrency,
-	type ThreadEvent,
-	type ThreadRun,
-	type ThreadRunHandle,
 	ThreadRuntime,
-	type ThreadRuntimeOptions,
-	type ThreadStartRunOptions,
-	type TreeSendOptions,
-	type TreeStateSnapshot,
 } from "./runtime";
+
+export type {
+	MessageTreeSnapshot,
+	ThreadConcurrency,
+	ThreadEvent,
+	ThreadRun,
+	ThreadRunHandle,
+	ThreadRuntimeOptions,
+	ThreadStartRunOptions,
+	ThreadStateSnapshot,
+	TreeSendOptions,
+} from "./types";
