+++ packages/thread/src/runtime.ts
@@ -1,8 +1,6 @@
 import {
-	AbstractChat,
-	type ChatInit,
+	type AbstractChat,
 	type ChatRequestOptions,
-	type ChatState,
 	type ChatStatus,
 	type ChatTransport,
 	type CreateUIMessage,
@@ -11,108 +9,29 @@ import {
 	generateId,
 	type UIMessage,
 } from "ai";
+import { ThreadRunChat } from "./thread-run-chat";
+import type {
+	MessageTreeSnapshot,
+	SendMessageInput,
+	ThreadConcurrency,
+	ThreadEvent,
+	ThreadRun,
+	ThreadRunHandle,
+	ThreadRunSpec,
+	ThreadRuntimeOptions,
+	ThreadStartRunOptions,
+	ThreadStateSnapshot,
+	TreeSendOptions,
+} from "./types";
 
 export const ROOT_PARENT_ID = "__root__";
-
-export type ThreadRun = {
-	assistantMessageId: string;
-	error: Error | undefined;
-	follow: boolean;
-	id: string;
-	originCursorId: string | null;
-	parentMessageId: string | null;
-	status: ChatStatus;
-	userMessageId: string;
-};
-
-export type ThreadRunHandle = {
-	readonly assistantMessageId: string;
-	readonly finished: Promise<void>;
-	readonly id: string;
-	getSnapshot: () => ThreadRun | undefined;
-	stop: () => Promise<void>;
-};
-
-export type TreeSendOptions = ChatRequestOptions & {
-	tree?: {
-		follow?: boolean;
-		from?: string | null;
-		[key: string]: unknown;
-	};
-};
-
-export type ThreadStartRunOptions<TMessage extends UIMessage = UIMessage> = {
-	follow?: boolean;
-	from?: string | null;
-	message?: SendMessageInput<TMessage>;
-	request?: TreeSendOptions;
-};
-
-export type ThreadConcurrency = {
-	maxActiveRuns?: number;
-	maxActiveRunsPerMessage?: number;
-};
-
-export type ThreadEvent =
-	| { cursorId: string | null; type: "cursor-changed" }
-	| { run: ThreadRun; type: "run-started" | "run-updated" }
-	| {
-			run: ThreadRun;
-			type: "run-aborted" | "run-completed" | "run-failed";
-	  };
-
-export type MessageTreeSnapshot<TMessage extends UIMessage = UIMessage> = {
-	childrenByParentId: Record<string, string[]>;
-	cursorId: string | null;
-	messagesById: Record<string, TMessage>;
-	parentById: Record<string, string | null>;
-	rootIds: string[];
-	version: 1;
-};
-
-export type TreeStateSnapshot<TMessage extends UIMessage = UIMessage> =
-	MessageTreeSnapshot<TMessage> & {
-		error: Error | undefined;
-		lastEvent: string;
-		messages: TMessage[];
-		status: ChatStatus;
-		activeRuns: ThreadRun[];
-		runs: ThreadRun[];
-		storeVersion: number;
-		treeStatus: ChatStatus;
-	};
-
-export type MessageTreeStore<TMessage extends UIMessage = UIMessage> = {
-	addMessage: (message: TMessage, parentId: string | null) => void;
-	getSnapshot: () => MessageTreeSnapshot<TMessage>;
-	getChildren: (messageId: string | null) => TMessage[];
-	getLeaves: (messageId?: string | null) => TMessage[];
-	getMessage: (messageId: string) => TMessage | undefined;
-	getParent: (messageId: string) => TMessage | undefined;
-	getPath: (messageId?: string | null) => TMessage[];
-	getSiblings: (messageId: string) => TMessage[];
-	replacePath: (messages: TMessage[]) => void;
-	setCursor: (messageId: string | null) => void;
-	setCursorToParentOf: (messageId: string) => void;
-	stopAll: () => Promise<void>;
-	stopRun: (runId: string) => Promise<void>;
-};
-
-type RunSpec = {
-	assistantMessageId: string;
-	follow: boolean;
-	originCursorId: string | null;
-	parentMessageId: string | null;
-	runId: string;
-	userMessageId: string;
-};
 
 type RunRecord<TMessage extends UIMessage> = {
 	aborted: boolean;
 	chat: ThreadRunChat<TMessage>;
 	error: Error | undefined;
 	finished: Promise<void>;
-	spec: RunSpec;
+	spec: ThreadRunSpec;
 	status: ChatStatus;
 };
 
@@ -122,20 +41,6 @@ function getAssistantMessageIdFromOptions(options?: TreeSendOptions) {
 	)?.assistantMessageId;
 	return typeof assistantMessageId === "string" ? assistantMessageId : null;
 }
-
-export type ThreadRuntimeOptions<TMessage extends UIMessage = UIMessage> = Omit<
-	ChatInit<TMessage>,
-	"messages"
-> & {
-	concurrency?: ThreadConcurrency;
-	initialTree?: MessageTreeSnapshot<TMessage>;
-	messages?: TMessage[];
-	onThreadEvent?: (event: ThreadEvent) => void;
-};
-
-export type SendMessageInput<TMessage extends UIMessage> = Parameters<
-	AbstractChat<TMessage>["sendMessage"]
->[0];
 
 function getInputMessageId<TMessage extends UIMessage>(
 	input: NonNullable<SendMessageInput<TMessage>>,
@@ -214,124 +119,6 @@ async function createUserMessageFromInput<TMessage extends UIMessage>({
 	} as TMessage;
 }
 
-class ThreadChatState<TMessage extends UIMessage>
-	implements ChatState<TMessage>
-{
-	#error: Error | undefined;
-	readonly #runtime: ThreadRuntime<TMessage>;
-	readonly #spec: RunSpec;
-	#status: ChatStatus = "ready";
-
-	constructor(runtime: ThreadRuntime<TMessage>, spec: RunSpec) {
-		this.#runtime = runtime;
-		this.#spec = spec;
-	}
-
-	get error() {
-		return this.#error;
-	}
-
-	set error(error: Error | undefined) {
-		this.#error = error;
-		this.#runtime.setRunError(this.#spec.runId, error);
-	}
-
-	get messages() {
-		const leafId = this.#runtime.hasAssistantStarted(
-			this.#spec.assistantMessageId,
-		)
-			? this.#spec.assistantMessageId
-			: this.#spec.userMessageId;
-		return this.#runtime.getPath(leafId);
-	}
-
-	set messages(messages: TMessage[]) {
-		this.#runtime.mergeRunPath(messages);
-	}
-
-	get status() {
-		return this.#status;
-	}
-
-	set status(status: ChatStatus) {
-		this.#status = status;
-		this.#runtime.setRunStatus(this.#spec.runId, status);
-	}
-
-	popMessage = () => {
-		const lastMessage = this.messages.at(-1);
-		if (lastMessage) {
-			this.#runtime.removeMessage(lastMessage.id);
-		}
-	};
-
-	pushMessage = (message: TMessage) => {
-		if (message.role === "assistant") {
-			const assistantMessage = {
-				...message,
-				id: this.#spec.assistantMessageId,
-			};
-			this.#runtime.markAssistantStarted(this.#spec.assistantMessageId);
-			this.#runtime.upsertMessage(assistantMessage, this.#spec.userMessageId);
-			this.#runtime.indexMessageOwnership(this.#spec.runId, assistantMessage);
-			return;
-		}
-
-		this.#runtime.upsertMessage(message, this.#spec.parentMessageId);
-	};
-
-	replaceMessage = (_index: number, message: TMessage) => {
-		if (message.role === "assistant") {
-			const assistantMessage = {
-				...message,
-				id: this.#spec.assistantMessageId,
-			};
-			this.#runtime.markAssistantStarted(this.#spec.assistantMessageId);
-			this.#runtime.upsertMessage(assistantMessage, this.#spec.userMessageId);
-			this.#runtime.indexMessageOwnership(this.#spec.runId, assistantMessage);
-			return;
-		}
-
-		this.#runtime.upsertMessage(message, this.#spec.parentMessageId);
-	};
-
-	snapshot = <T>(thing: T): T => clone(thing);
-}
-
-class ThreadRunChat<TMessage extends UIMessage> extends AbstractChat<TMessage> {
-	constructor(runtime: ThreadRuntime<TMessage>, spec: RunSpec) {
-		super({
-			dataPartSchemas: runtime.dataPartSchemas,
-			generateId: () => spec.assistantMessageId,
-			id: runtime.chatId,
-			messageMetadataSchema: runtime.messageMetadataSchema,
-			onData: (event) => runtime.onData?.(event),
-			onError: (error) => {
-				runtime.setRunError(spec.runId, error);
-				runtime.onError?.(error);
-			},
-			onFinish: (event) => {
-				const assistantMessage = {
-					...event.message,
-					id: spec.assistantMessageId,
-				};
-				runtime.upsertMessage(assistantMessage, spec.userMessageId);
-				runtime.indexMessageOwnership(spec.runId, assistantMessage);
-				runtime.finishRequest(spec.runId, event.isAbort);
-				runtime.onFinish?.({ ...event, message: assistantMessage });
-			},
-			onToolCall: async (event) => {
-				runtime.registerToolCall(spec.runId, event.toolCall.toolCallId);
-				await runtime.onToolCall?.(event);
-			},
-			sendAutomaticallyWhen: (event) =>
-				runtime.sendAutomaticallyWhen?.(event) ?? false,
-			state: new ThreadChatState(runtime, spec),
-			transport: runtime.transport,
-		});
-	}
-}
-
 export class ThreadRuntime<TMessage extends UIMessage = UIMessage> {
 	readonly chatId: string;
 	readonly dataPartSchemas: ThreadRuntimeOptions<TMessage>["dataPartSchemas"];
@@ -345,7 +132,7 @@ export class ThreadRuntime<TMessage extends UIMessage = UIMessage> {
 	onThreadEvent: ThreadRuntimeOptions<TMessage>["onThreadEvent"];
 	onToolCall: ThreadRuntimeOptions<TMessage>["onToolCall"];
 	sendAutomaticallyWhen: ThreadRuntimeOptions<TMessage>["sendAutomaticallyWhen"];
-	readonly transport: ChatTransport<TMessage>;
+	transport: ChatTransport<TMessage>;
 
 	readonly #assistantStarted = new Set<string>();
 	readonly #childrenByParentId = new Map<string, string[]>();
@@ -358,7 +145,7 @@ export class ThreadRuntime<TMessage extends UIMessage = UIMessage> {
 	readonly #runsById = new Map<string, RunRecord<TMessage>>();
 	#cursorId: string | null = null;
 	#lastEvent = "Ready";
-	#snapshot: TreeStateSnapshot<TMessage>;
+	#snapshot: ThreadStateSnapshot<TMessage>;
 	#storeVersion = 0;
 
 	constructor(options: ThreadRuntimeOptions<TMessage> = {}) {
@@ -392,13 +179,16 @@ export class ThreadRuntime<TMessage extends UIMessage = UIMessage> {
 
 	getSnapshot = () => this.#snapshot;
 
-	updateCallbacks(options: ThreadRuntimeOptions<TMessage>) {
+	updateOptions(options: ThreadRuntimeOptions<TMessage>) {
 		this.onData = options.onData;
 		this.onError = options.onError;
 		this.onFinish = options.onFinish;
 		this.onThreadEvent = options.onThreadEvent;
 		this.onToolCall = options.onToolCall;
 		this.sendAutomaticallyWhen = options.sendAutomaticallyWhen;
+		if (options.transport) {
+			this.transport = options.transport;
+		}
 	}
 
 	subscribe = (listener: () => void) => {
@@ -827,7 +617,7 @@ export class ThreadRuntime<TMessage extends UIMessage = UIMessage> {
 		}
 	}
 
-	private buildSnapshot(): TreeStateSnapshot<TMessage> {
+	private buildSnapshot(): ThreadStateSnapshot<TMessage> {
 		const messagesById = Object.fromEntries(
 			Array.from(this.#messagesById.entries()).map(([id, message]) => [
 				id,
@@ -890,7 +680,7 @@ export class ThreadRuntime<TMessage extends UIMessage = UIMessage> {
 		return true;
 	}
 
-	private createStreamChat(spec: RunSpec) {
+	private createStreamChat(spec: ThreadRunSpec) {
 		return new ThreadRunChat(this, spec);
 	}
 
@@ -906,7 +696,7 @@ export class ThreadRuntime<TMessage extends UIMessage = UIMessage> {
 		if (!userMessageId) {
 			return undefined;
 		}
-		const spec: RunSpec = {
+		const spec: ThreadRunSpec = {
 			assistantMessageId: assistantMessage.id,
 			follow: true,
 			originCursorId: assistantMessage.id,
@@ -1107,7 +897,7 @@ export class ThreadRuntime<TMessage extends UIMessage = UIMessage> {
 	}
 
 	private startRunRequest(
-		specWithOptions: RunSpec & { options?: TreeSendOptions },
+		specWithOptions: ThreadRunSpec & { options?: TreeSendOptions },
 	): ThreadRunHandle {
 		const { options, ...spec } = specWithOptions;
 		const chat = this.createStreamChat(spec);
@@ -1185,7 +975,7 @@ export class ThreadRuntime<TMessage extends UIMessage = UIMessage> {
 	}
 
 	private withRunRequestBody(
-		spec: RunSpec,
+		spec: ThreadRunSpec,
 		options: ChatRequestOptions = {},
 	): ChatRequestOptions {
 		const treeOptions =
