+++ packages/thread/tsconfig.json
@@ -12,5 +12,5 @@
 		"isolatedModules": true,
 		"types": ["bun"]
 	},
-	"include": ["src/**/*.ts", "spike/**/*.ts", "test/**/*.ts"]
+	"include": ["src/**/*.ts", "test/**/*.ts"]
 }
